feat(mapgen-studio): decompose App.tsx into StudioProviders/StudioShell + chrome tree
Split the remaining ~2,583-LoC App.tsx god-component into the component tree per
architecture/10 §4 + audit/03. The AppContent orchestration closure MOVES verbatim
into src/app/StudioShell.tsx; the provider wrapper, deck.gl canvas host, floating
docks, error banner, and sonner toast adapter become standalone units under
src/app/. App.tsx reduces to a 17-LoC thin root rendering StudioProviders.

Behavior-preserving by construction: state web, effects, refs, and all hard-core
behaviors (live-poll request-key staleness + adaptive backoff, run-in-game
fingerprint/relation/materialization, browserRunner gating, localStorage schema)
move byte-for-byte. Live reads still flow through the typed oRPC client; no new
hand-rolled fetch. Verified: bun run check (tsc clean), bun run build (incl
worker-bundle), live preview (recipe authoring + stage/step nav + run controls
present and functional, no console errors).

OpenSpec: mapgen-studio-app-shell (validated --strict).
