+++ openspec/changes/mapgen-studio-app-shell/.openspec.yaml
@@ -1,0 +1,3 @@
+schema: spec-driven
+created: 2026-06-09
+status: draft
